% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoarseGraining}
\alias{CoarseGraining}
\title{Signal Scaling (coarse-grained by average)}
\usage{
CoarseGraining(x, scale)
}
\arguments{
\item{x}{the signal as a numeric vector}

\item{scale}{the scaling factor}
}
\value{
a numeric vector with the coarse-grained signal
}
\description{
Compute a coarse-grained copy of an input signal. The signal is average on
non-overlapping windows of size 'scale' (the scale factor).
}
