% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisEn.R
\name{DispersionEntropy}
\alias{DispersionEntropy}
\title{Dispersion Entropy}
\usage{
DispersionEntropy(x, ma = "NCDF", m = 3, nc = 6, tau = 1, mu,
  sigma)
}
\arguments{
\item{x}{an univariate signal (vector)}

\item{ma}{the mapping approach, possible values are:
\itemize{
   \item \code{LM}: linear mapping
   \item \code{NCDF}: normal cumulative distribution function (default)
   \item \code{LOGSIG}: logarithm sigmoid
   \item \code{TANSIG}: tangent sigmoid
   \item \code{SORT}: sorting method
}}

\item{m}{the embedding dimension}

\item{nc}{the number of classes (it is usually equal to a number between 3 and 9, we use 6 by default)}

\item{tau}{the time lag (it is usually equal to 1)}

\item{mu}{an optional defined mean (the observed mean by default)}

\item{sigma}{an option defined std-deviation (the observed std-deviation by default)}
}
\value{
a named list:
\itemize{
   \item \code{disp.en}: the Dispersion Entropy value of the signal x
   \item \code{pdf}: a vector of length nc^m, showing the normalized number of disersion patterns of x
}
}
\description{
This function calculates dispersion entropy of a univariate signal
x, using different mapping approaches (MA)
}
\examples{
data(EG_181117)
DispersionEntropy(EG_181117,m=1)
}
\references{
\enumerate{
   \item Azami H, Entropy JE, (2018) "Amplitude-and Fluctuation-Based Dispersion Entropy". Entropy 20:210.
   \item Rostaghi M, Letters HAISP, (2016) "Dispersion entropy: A measure for time-series analysis". IEEE Signal Processing Letters. vol. 23, n. 5, pp. 610-614, 2016.
}
}
