% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_181117_2.R
\docType{data}
\name{EG_181117_2}
\alias{EG_181117_2}
\title{EEG Dataset 2}
\format{An object of class \code{numeric} of length 10640.}
\usage{
data(EG_181117_2)
}
\description{
To be completed
}
\keyword{datasets}
