% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_181218.R
\docType{data}
\name{EG_181218}
\alias{EG_181218}
\title{EG_181218}
\format{An object of class \code{numeric} of length 6857.}
\usage{
data(EG_181218)
}
\description{
R-R measurements of a 41 Male with no particular pathology during a moderate
running training. Measurements were caputered using a Polar H10 cardio belt
and processed with the HRV+ iPhone application.
}
\keyword{datasets}
