% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{Multi Scale Entropy (MSE)}
\usage{
MSE(x, m = 2, r = 0.15, scales = 1:10)
}
\arguments{
\item{x}{a univariate signal (vector)}

\item{m}{the embedding dimension (default: 2)}

\item{r}{the tolerance factor for "similarity" (default: 0.15)}

\item{scales}{the scale factors as a vector (default 1:10)}
}
\value{
a vector of size 1 * length(scales) - the RCMDE of x
}
\description{
This function calculates the multiscale entropy (MSE) of a univariate signal x
}
\examples{
data(EG_181117)
MSE(EG_181117)
}
\references{
\enumerate{
   \item  Costa M, Goldberger AL, Peng CK (2005) "Multiscale entropy analysis of biological signals". Phys. Rev. E - Stat. Nonlinear, Soft Matter Phys. 71.
}
}
