% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{randomSignal}
\alias{randomSignal}
\title{Generate a random signal}
\usage{
randomSignal(n, model = "white", alpha = 1)
}
\arguments{
\item{n}{the length of the signal to generate}

\item{model}{the type of signal to generate}

\item{alpha}{The power for the power law noise model
(defaults are 1 for pink and 1.5 for red noise) \eqn{(1/f)^alpha}.}
}
\value{
a vector
}
\description{
This function generate a random signal using the noise function of the
`tuneR` package.
}
\details{
Signal generated by these function can be built using different models. By
default this is a \emph{white noise}, but one can also used
\emph{pink, red or power} models.
}
\examples{
x <- randomSignal(1000)
\donttest{plot(x,type='l')}
}
