% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampEn}
\alias{SampEn}
\title{Sample Entropy}
\usage{
SampEn(x, m, r, sd)
}
\arguments{
\item{x}{the signal as a numeric vector}

\item{m}{the embedding dimension}

\item{r}{the tolerance factor for "similarity"}

\item{sd}{the standard-deviation of the signal}
}
\value{
the sample entropy value
}
\description{
Calcul the sample entropy
}
