% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_181117_2.R
\docType{data}
\name{EG_181117_2}
\alias{EG_181117_2}
\title{EG_181117_2}
\format{
An object of class \code{numeric} of length 10640.
}
\usage{
data(EG_181117_2)
}
\description{
R-R measurements of a 41 Male with no particular pathology in rest position.
Measurements were obtained using a Polar RS800 running computer.
}
\keyword{datasets}
