% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{randomSignal}
\alias{randomSignal}
\title{Generate a random signal}
\usage{
randomSignal(n)
}
\arguments{
\item{n}{the length of the signal to generate}
}
\value{
a vector
}
\description{
This function generate a random signal using a very simple random walk
}
\examples{
x <- randomSignal(1000)
\donttest{plot(x,type='l')}
}
