% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDE.R
\name{RCMDE}
\alias{RCMDE}
\title{Refined Composite Multiscale Dispersion Entropy}
\usage{
RCMDE(x, m = 1, nc = 6, tau = 1, scales = 1:10)
}
\arguments{
\item{x}{a univariate signal (vector)}

\item{m}{the embedding dimension (default: 1)}

\item{nc}{the number of classes (it is usually equal to a number between 3 and 9, we use 6 by default)}

\item{tau}{the time lag (it is usually equal to 1)}

\item{scales}{the scale factors as a vector (default 1:10)}
}
\value{
a vector of size 1 * length(scales) - the RCMDE of x
}
\description{
This function calculates the refined composite multiscale dispersion entropy (RCMDE) of a univariate signal x
}
\examples{
data(EG_181117)
RCMDE(EG_181117)
}
\references{
\enumerate{
   \item  Azami H, Rostaghi M, Abasolo D, Escudero J (2017) "Refined Composite Multiscale Dispersion Entropy and its Application to Biomedical Signals". IEEE transactions on bio-medical engineering 64:2872–2879.
   \item Rostaghi M, Letters HAISP, (2016) "Dispersion entropy: A measure for time-series analysis". IEEE Signal Processing Letters. vol. 23, n. 5, pp. 610-614, 2016.
}
}
